use foldseek webserver if local args not provided

+++ plaseek/tools/foldseek.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from logging import StreamHandler, Formatter
 import os
 import subprocess
 from typing import Union
@@ -8,16 +7,7 @@ from time import sleep
 import logging
 import tarfile
 
-
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
-stream_handler = StreamHandler()
-stream_handler.setLevel(logging.DEBUG)
-handler_format = Formatter(
-    "%(asctime)s %(filename)s:%(lineno)d - %(levelname)s - %(message)s"
-)
-stream_handler.setFormatter(handler_format)
-logger.addHandler(stream_handler)
 
 
 def run_foldseek_locally(
@@ -63,65 +53,75 @@ def run_foldseek_locally(
 
 
 def retrieve_foldseek_results(
-    pdbfile: str | Path, mode="3diaa", database: list = ["afdb50", "mgnify_esm30"]
+    pdbfile: str | Path,
+    output_directory: str | Path,
+    mode="3diaa",
+    database: list = ["afdb50", "mgnify_esm30"],
 ) -> tuple[Path, list[str]]:
     """Retrieve Foldseek results from the webserver.
     Args:
         pdbfile: Path to input pdb file.
         m8file: Path to output m8 file.
-        mode: Mode of Foldseek.
+        mode: Mode of Foldseek. "3diaa" or "tmalign".
         database: Database of Foldseek.
     Returns:
 
     """
     # reference : https://github.com/soedinglab/MMseqs2-App/blob/master/docs/api_example.py
+    if mode not in ["3diaa", "tmalign"]:
+        raise ValueError(f"mode should be 3diaa or tmalign, but got {mode}.")
     if not Path(pdbfile).exists():
         raise FileNotFoundError(f"{pdbfile} not found.")
     file_name = Path(pdbfile).stem
+    output_directory = Path(output_directory)
+    result_file = output_directory.joinpath(f"result_{file_name}.tar.gz")
+    if result_file.exists():
+        logger.info(f"{result_file} already exists. Skip running Foldseek.")
+    else:
+        files = {"q": open(pdbfile, "rb")}
+        data = {
+            "mode": mode,  # select the mode you want to use
+            "database[]": database,  # select the database you want to use
+        }
 
-    files = {"q": open(pdbfile, "rb")}
-    data = {
-        "mode": mode,  # select the mode you want to use
-        "database[]": database,  # select the database you want to use
-    }
+        ticket = post(
+            "https://search.foldseek.com/api/ticket", files=files, data=data
+        ).json()
 
-    ticket = post(
-        "https://search.foldseek.com/api/ticket", files=files, data=data
-    ).json()
+        print(ticket)
 
-    print(ticket)
+        # poll until the job was successful or failed
+        repeat = True
+        while repeat:
+            status = get(
+                "https://search.foldseek.com/api/ticket/" + ticket["id"]
+            ).json()
+            logger.info(f"current status is {status['status']}")
+            if status["status"] == "ERROR":
+                raise RuntimeError(f"Foldseek failed with error: {status['error']}")
 
-    # poll until the job was successful or failed
-    repeat = True
-    while repeat:
-        status = get("https://search.foldseek.com/api/ticket/" + ticket["id"]).json()
-        logger.info(f"current status is {status['status']}")
-        if status["status"] == "ERROR":
-            raise RuntimeError(f"Foldseek failed with error: {status['error']}")
+            # wait a short time between poll requests
+            sleep(5)
+            # if status['status'] is not "COMPLETE" repeat = True
+            repeat = status["status"] != "COMPLETE"
 
-        # wait a short time between poll requests
-        sleep(5)
-        # if status['status'] is not "COMPLETE" repeat = True
-        repeat = status["status"] != "COMPLETE"
+        download = get(
+            "https://search.foldseek.com/api/result/download/" + ticket["id"],
+            stream=True,
+        )
 
-    download = get(
-        "https://search.foldseek.com/api/result/download/" + ticket["id"], stream=True
-    )
+        with open(result_file, "wb") as fd:
+            for chunk in download.iter_content(chunk_size=128):
+                fd.write(chunk)
 
-    result_file = Path(f"result_{file_name}.tar.gz")
-    with open(result_file, "wb") as fd:
-        for chunk in download.iter_content(chunk_size=128):
-            fd.write(chunk)
     return result_file, database
 
 
-def write_merged_m8file(
-    result_file: Path, mergedm8file: str = "alis_merged.m8"
-) -> Path:
-    """Merge m8 files in the {result_file}.tar.gz."""
+def write_merged_m8file(resulttargzfile: Path, mergedm8file: Path) -> Path:
+    """Merge m8 files in the {result_file}"""
     with open(mergedm8file, "w") as outfile:
-        with tarfile.open(result_file, "r") as tarf:
-            logger.info(f"Extracting m8 files from {result_file}.tar.gz")
+        with tarfile.open(resulttargzfile, "r") as tarf:
+            logger.info(f"Extracting m8 files from {resulttargzfile}.")
             for member in tarf.getmembers():
                 if member.name.endswith(".m8"):
                     logger.info(f"Found {member.name}")
@@ -129,3 +129,26 @@ def write_merged_m8file(
                     if f is not None:
                         outfile.write(f.read().decode())
     return Path(mergedm8file)
+
+
+def run_foldseek_webserver(
+    input: Path,
+    output_directory: Path,
+    mode="3diaa",
+    database=["afdb50", "mgnify_esm30"],
+) -> Path:
+    """run Foldseek webserver and return the path to the merged m8 file.
+    Args:
+        input: Path to input pdb file.
+        mode: Mode of Foldseek. "3diaa" or "tmalign".
+        database: Database of Foldseek.
+    """
+    assert Path(input).exists(), f"{input} not found."
+    assert input.suffix == ".pdb", f"input file should be .pdb, but got {input.suffix}."
+    resulttargzfile, database = retrieve_foldseek_results(
+        input, output_directory, mode, database
+    )
+    mergedm8file = write_merged_m8file(
+        resulttargzfile, Path(output_directory.joinpath(f"{input.stem}.m8"))
+    )
+    return mergedm8file
